requirements.txt
# 图像处理和计算机视觉
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3

# OCR 识别
pytesseract==0.3.10
easyocr==1.7.0

# 深度学习和YOLO - CUDA 12.1版本
torch==2.5.1+cu121
torchvision==0.20.1+cu121
torchaudio==2.5.1+cu121
ultralytics==8.3.0
onnxruntime-gpu==1.19.2

# 图像增强和预处理
scikit-image==0.21.0
imutils==0.5.4

# 数据处理
pandas==2.1.1
matplotlib==3.7.2
seaborn==0.12.2

# 工具库
tqdm==4.66.1
loguru==0.7.2
pyyaml==6.0.1
requests==2.31.0

# 额外的PyTorch相关包
--extra-index-url https://download.pytorch.org/whl/cu121 